Post feed scaled
Post feed page is scaled to different resolutions.

Messages in consoel.log is removed, but a counter is placed (counting total posts)

every post is linking to google.com (will be changed to single post page when created.

added a mobile navigationbar that will show untill width 768px is reached.