finished create-post page, made deleting posts possible and started on single-post page
The createPost page is almost completed. Some styling and error-message displays are needed if an error occours.

on the profile page will you now be able to see your own posts and be able to delete them. more styling is needed.

single post page is under construction. profileData and API-key is retrieved but needs to be used to get the single post. Still working on it.