small changes and new page
Edit post page is now retrieving information from a post but not displaying it on the screen yet but in the console.log().

changed storage saving from local to session for profileData.

made author visible on feedpage. now we can see the author's username on each posts with a a-tag that later will lead to the profile page for that user.

will soon work with searching up posts but will try to add users to it as well.