finished see user profile page
The see user profile page is finished adn wortking good. Next thing is work a little more on styling the page with other pages.

Clicking on the post will send you to the single-post page. This connection goes trhough your own profile page and the profile page to the user you are watching.