updating and bug-fixing
error: when you click on profile in the nav bar when you are on the see profile page, you were sent to the landing page

error is fixed now.

added a kickout function to all pages:
if no user is detected, you will be sent back to the landing page and will have to sign in again.

added description for use of this site.

mad it possible to update profile (see user named marTest)

removed some console.logs for pages where they were not needed no more.